refactor: removes async fs calls
